refactor: [#31] move function to production code
